+++ packages/stream_chat/lib/src/client/client.dart
@@ -70,8 +70,15 @@ class StreamChatClient {
     Duration receiveTimeout = const Duration(seconds: 6),
     StreamChatApi? chatApi,
     WebSocket? ws,
-    AttachmentFileUploader? attachmentFileUploader,
-  }) {
+    @Deprecated('Use [attachmentFileUploaderProvider] instead')
+        AttachmentFileUploader? attachmentFileUploader,
+    AttachmentFileUploaderProvider? attachmentFileUploaderProvider,
+  }) : assert(
+          attachmentFileUploader == null ||
+              attachmentFileUploaderProvider == null,
+          'You can only use one of [attachmentFileUploader] '
+          'or [attachmentFileUploaderProvider]',
+        ) {
     logger.info('Initiating new StreamChatClient');
 
     final options = StreamHttpClientOptions(
@@ -81,13 +88,23 @@ class StreamChatClient {
       headers: {'X-Stream-Client': defaultUserAgent},
     );
 
+    // TODO: simplify this once we remove the deprecated field.
+    final AttachmentFileUploaderProvider fileUploaderProvider;
+    if (attachmentFileUploaderProvider != null) {
+      fileUploaderProvider = attachmentFileUploaderProvider;
+    } else if (attachmentFileUploader != null) {
+      fileUploaderProvider = (httpClient) => attachmentFileUploader;
+    } else {
+      fileUploaderProvider = StreamAttachmentFileUploader.new;
+    }
+
     _chatApi = chatApi ??
         StreamChatApi(
           apiKey,
           options: options,
           tokenManager: _tokenManager,
           connectionIdManager: _connectionIdManager,
-          attachmentFileUploader: attachmentFileUploader,
+          attachmentFileUploaderProvider: fileUploaderProvider,
           logger: detachedLogger('🕸️'),
         );