+++ packages/stream_chat/CHANGELOG.md
@@ -4,6 +4,15 @@
 
 - Fix WebSocket contemporary connection calls while disconnecting
 
+✅ Added
+
+- Export `StreamAttachmentFileUploader`.
+
+🔄 Changed
+
+- Deprecated `StreamChatClient.attachmentFileUploader`,
+  Use `StreamChatClient.attachmentFileUploaderProvider` instead.
+
 ## 4.3.0
 
 🐞 Fixed